[FEATURE] 휴게소 상세 정보 완료

+++ core/network/src/main/java/com/eight_potato/network/model/request/RouteRequest.kt
@@ -7,13 +7,13 @@ import com.google.gson.annotations.SerializedName
  */
 data class RouteRequest(
     @SerializedName("startX")
-    val startX: Float, // 출발지 경도
+    val startX: Double, // 출발지 경도
     @SerializedName("startY")
-    val startY: Float, // 출발지 위도
+    val startY: Double, // 출발지 위도
     @SerializedName("endX")
-    val endX: Float, // 도착지 경도
+    val endX: Double, // 도착지 경도
     @SerializedName("endY")
-    val endY: Float, // 도착지 위도
+    val endY: Double, // 도착지 위도
     @SerializedName("reqCoordType")
     val reqCoordType: String = "WGS84GEO",
     @SerializedName("resCoordType")
